Modify Renesas RL78/G14 FPB {CC-RX, GCC} RTOS full demo a little bit to be similar to Renesas RX72N Envision Kit RTOS full demo

+++ FreeRTOS/Demo/Common/Minimal/TaskNotify.c
@@ -106,7 +106,7 @@ static uint32_t ulTimerNotificationsReceived = 0UL, ulTimerNotificationsSent = 0
 static TimerHandle_t xTimer = NULL;
 
 /* Used by the pseudo random number generating function. */
-static size_t uxNextRand = 0;
+static uint32_t uxNextRand = 0;
 
 /*-----------------------------------------------------------*/
 
@@ -723,7 +723,7 @@ const uint32_t ulMaxSendReceiveDeviation = 5UL;
 
 static UBaseType_t prvRand( void )
 {
-const size_t uxMultiplier = ( size_t ) 0x015a4e35, uxIncrement = ( size_t ) 1;
+const uint32_t uxMultiplier = ( uint32_t ) 0x015a4e35, uxIncrement = ( uint32_t ) 1;
 
 	/* Utility function to generate a pseudo random number. */
 	uxNextRand = ( uxMultiplier * uxNextRand ) + uxIncrement;
